Add OAuth2Manager and related tests
- Introduced `OAuth2Manager` to manage OAuth2 authentication flow, including token retrieval, refresh, and error handling.
- Implemented `FileTokenStorage` for managing OAuth2 token data using file storage, with methods for saving, loading, and clearing tokens.
- Added comprehensive unit tests for `OAuth2Manager`, covering scenarios for token retrieval, refresh, and error handling.
- Enhanced type definitions for `TokenData`, `TokenStorageProvider`, and `OAuth2Config` to improve type safety and clarity.

These changes establish a robust foundation for OAuth2 integration, enhancing token management and authentication processes in the application.
