+++ workers/main/package.json
@@ -11,6 +11,7 @@
     "@eslint/js": "9.27.0",
     "@temporalio/testing": "1.11.8",
     "@types/node": "22.15.21",
+    "@types/simple-oauth2": "5.0.7",
     "@vitest/coverage-v8": "3.1.3",
     "c8": "10.1.3",
     "dotenv": "16.5.0",
@@ -40,6 +41,7 @@
     "axios-retry": "4.5.0",
     "mongoose": "8.15.1",
     "mysql2": "3.14.1",
+    "simple-oauth2": "5.1.0",
     "zod": "3.25.17"
   }
 }
